requirements.txt
dspy-ai>=2.4.0
langgraph>=0.2.30
langchain-core>=0.2.0
pydantic>=2.0.0
click>=8.1.7
rich>=13.7.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2
matplotlib>=3.7.0
networkx>=3.1
graphviz>=0.20.1
seaborn>=0.12.0
plotly>=5.15.0
ipywidgets>=8.0.0
pyppeteer>=2.0.0
